requirements.txt
requests
loguru
python-decouple
pandas
openpyxl